scripts/requirements_pipeline.txt
# Data Processing Pipeline Dependencies

# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Database connections
psycopg2-binary>=2.9.0
chromadb>=0.4.0

# File processing
PyPDF2>=3.0.0
openpyxl>=3.0.0
xlrd>=2.0.0

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.0.0

# Configuration and utilities
PyYAML>=6.0
python-dotenv>=0.19.0

# Data validation and cleaning
pydantic>=1.10.0
jsonschema>=4.0.0

# Logging and monitoring
structlog>=22.0.0

# Optional: For advanced data processing
# scikit-learn>=1.1.0
# nltk>=3.7
# spacy>=3.4.0

# Optional: For web scraping with JavaScript rendering
# playwright>=1.25.0

# Optional: For rate limiting and proxy management
# aiohttp>=3.8.0
# asyncio-throttle>=1.0.0
